phase4-ml/serving/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mlflow==2.10.0
scikit-learn==1.4.0
numpy==1.26.4
redis==5.0.1
psycopg2-binary==2.9.9
boto3==1.34.0
